requirements.txt
numpy~=1.23.5
pandas~=2.0.0
torch~=1.13.1
torchvision~=0.15.0
pillow~=9.4.0
einops~=0.6.0
matplotlib~=3.7.1
tqdm~=4.65.0
scikit-learn~=1.2.2
opencv-python~=4.7.0.72
imageio~=2.27.0
simpleitk~=2.2.1
scikit-image~=0.20.0